Add cucumber in Mobile

+++ core/src/main/java/common/Waiter.java
@@ -2,7 +2,7 @@ package common;
 
 public class Waiter {
 
-    public static void waitMS(int ms) {
+    public static void waitMS(long ms) {
         try {
             Thread.sleep(ms);
         } catch(InterruptedException ex) {
